feat(*): calculate entities per hour metric

+++ core/syncInfo.go
@@ -3,7 +3,7 @@ package core
 type SyncInfo struct {
     Index           int
     SyncedEntities  int
-    EntitiesPerHour float32
+    EntitiesPerHour int
     Target          string
     TaskId          int
     TaskStarted     string
